# apps/web/worker/routes/reports.routes.ts
import { Hono } from "hono";
import type { AppContext } from "../env";
import { requireAuth } from "../middleware/auth.middleware";
import { badRequest } from "../http/errors";
import { parseListLimit, parseListOffset } from "../http/params";
import { getBalanceSheet, getGeneralLedger, getProfitLoss } from "../services/reports.service";

export const reportsRoutes = new Hono<AppContext>();

reportsRoutes.use("*", requireAuth());

reportsRoutes.get("/profit-loss", async (c) => {
  const url = new URL(c.req.url);
  const report = await getProfitLoss(
    c.env.DB,
    c.get("user").id,
    requiredParam(url.searchParams, "fromDate"),
    requiredParam(url.searchParams, "toDate"),
  );
  c.res.headers.set("Cache-Control", "private, max-age=30");
  return c.json({ report });
});

reportsRoutes.get("/balance-sheet", async (c) => {
  const url = new URL(c.req.url);
  const report = await getBalanceSheet(
    c.env.DB,
    c.get("user").id,
    requiredParam(url.searchParams, "asOfDate"),
  );
  c.res.headers.set("Cache-Control", "private, max-age=30");
  return c.json({ report });
});

reportsRoutes.get("/general-ledger", async (c) => {
  const url = new URL(c.req.url);
  const params = url.searchParams;
  const report = await getGeneralLedger(
    c.env.DB,
    c.get("user").id,
    {
      accountId: params.get("accountId") || undefined,
      fromDate: requiredParam(params, "fromDate"),
      toDate: requiredParam(params, "toDate"),
      limit: parseListLimit(params.get("limit"), 5000),
      offset: parseListOffset(params.get("offset")),
    },
  );
  c.res.headers.set("Cache-Control", "private, max-age=30");
  return c.json({ report });
});

function requiredParam(params: URLSearchParams, name: string): string {
  const value = params.get(name);
  if (!value) throw badRequest("missing_query_param", `${name} is required`);
  return value;
}