# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ledjer API
  version: 1.0.0
  description: >
    Cloudflare Worker (Hono) API for Ledjer - cash-only double-entry
    bookkeeping MVP for Indonesian UMKM. All money values are integer IDR
    (minor units). Protected endpoints require the session cookie
    (`__Host-ledjer_session` in production, `ledjer_session` elsewhere) and
    state-changing requests must send a same-origin `Origin` header.
servers:
  - url: https://ledjer.id
    description: Production
  - url: http://localhost:8788
    description: "Local dev (via Vite proxy: http://localhost:5173)"
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a user, creating their book + default COA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName, businessName]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 72 }
                fullName: { type: string, minLength: 2, maxLength: 160 }
                businessName: { type: string, minLength: 1, maxLength: 120 }
      responses:
        "200":
          description: Registered; session cookie set
        "403":
          description: Email already taken
        "429":
          description: Rate limited
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in and set the session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Logged in; session cookie set
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
        "401":
          description: Invalid credentials
        "429":
          description: Rate limited
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user + session (null when unauthenticated)
      responses:
        "200":
          description: Current session state
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: [object, "null"]
                    properties:
                      id: { type: string }
                      email: { type: string }
                      full_name: { type: string }
                      business_name: { type: string }
                  session:
                    type: [object, "null"]
                    properties:
                      id: { type: string }
                      user_id: { type: string }
                      expires_at: { type: integer }
    patch:
      tags: [Auth]
      security: [{ cookieAuth: [] }]
      summary: Rename the book (business_name on the user profile)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [businessName]
              properties:
                businessName: { type: string, minLength: 1, maxLength: 120 }
      responses:
        "200":
          description: Updated business name
        "401":
          description: Not authenticated
  /api/auth/google/start:
    get:
      tags: [Auth]
      summary: Return a Google OAuth authorization URL (state cookie set)
      responses:
        "200":
          description: Authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
        "501":
          description: OAuth not configured
  /api/auth/google/callback:
    get:
      tags: [Auth]
      summary: OAuth callback - exchanges code, sets session, redirects to /dashboard
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        "302":
          description: Redirect (dashboard on success, /login?error=... on failure)
...
  /api/reports/profit-loss:
    get:
      tags: [Reports]
      security: [{ cookieAuth: [] }]
      summary: Profit & loss within a date range
      parameters:
        - { name: fromDate, in: query, required: true, schema: { type: string, format: date } }
        - { name: toDate, in: query, required: true, schema: { type: string, format: date } }
      responses:
        "200":
          description: Profit-loss report
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: object
                    properties:
                      fromDate: { type: string }
                      toDate: { type: string }
                      income: { $ref: "#/components/schemas/ReportTotal" }
                      expense: { $ref: "#/components/schemas/ReportTotal" }
                      netIncome: { type: integer }
  /api/reports/balance-sheet:
    get:
      tags: [Reports]
      security: [{ cookieAuth: [] }]
      summary: Balance sheet as of a date
      parameters:
        - { name: asOfDate, in: query, required: true, schema: { type: string, format: date } }
      responses:
        "200":
          description: Balance sheet
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: object
                    properties:
                      asOfDate: { type: string }
                      assets: { type: array, items: { $ref: "#/components/schemas/ReportLine" } }
                      totalAssets: { type: integer }
                      liabilities: { type: array, items: { $ref: "#/components/schemas/ReportLine" } }
                      totalLiabilities: { type: integer }
                      equity: { type: array, items: { $ref: "#/components/schemas/ReportLine" } }
                      totalEquity: { type: integer }
                      balanced: { type: boolean }
  /api/reports/general-ledger:
    get:
      tags: [Reports]
      security: [{ cookieAuth: [] }]
      summary: General ledger - journal lines per account with running balance
      parameters:
        - { name: fromDate, in: query, required: true, schema: { type: string, format: date } }
        - { name: toDate, in: query, required: true, schema: { type: string, format: date } }
        - { name: accountId, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 5000, default: 5000 } }
        - { name: offset, in: query, required: false, schema: { type: integer, minimum: 0 } }
      responses:
        "200":
          description: Ledger report
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: object
                    properties:
                      fromDate: { type: string }
                      toDate: { type: string }
                      accountId: { type: [string, "null"] }
                      truncated: { type: boolean }
                      entries:
                        type: array
                        items:
                          type: object
                          properties:
                            account_id: { type: string }
                            account_code: { type: string }
                            account_name: { type: string }
                            account_class:
                              type: string
                              enum: [asset, liability, equity, income, expense]
                            entry_date: { type: string, format: date }
                            transaction_id: { type: string }
                            transaction_number: { type: string }
                            description: { type: string }
                            debit_idr: { type: integer }
                            credit_idr: { type: integer }
                            running_balance_idr: { type: integer }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-ledjer_session
  schemas:
...
    ReportLine:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        amount: { type: integer }
    ReportTotal:
      type: object
      properties:
        total: { type: integer }
        accounts:
          type: array
          items: { $ref: "#/components/schemas/ReportLine" }